docs(viewer-docs): apply Annotation to Polygon Viewer, Drawer
