+++ apps/insight-viewer-docs/src/containers/UseOverlayContext/PolygonDrawer/index.tsx
@@ -1,19 +1,16 @@
 /* eslint-disable import/no-unresolved */
 import React from 'react'
-import { Box } from '@chakra-ui/react'
+import { Box, Button } from '@chakra-ui/react'
 import { Resizable } from 're-resizable'
 import InsightViewer, {
   Contour,
   useImage,
   useContour,
   useViewport,
-  SvgContourViewer,
+  Annotation,
 } from '@lunit/insight-viewer'
 import { IMAGES } from '../../../const'
-import { CONTOURS } from '../../../../mocks/contours'
 import { getPolygonStyles } from '../../../utils/common/getPolygonStyles'
-import CodeBlock from '../../../components/CodeBlock'
-import { CODE } from './Code'
 
 const style = {
   display: 'flex',
@@ -22,20 +19,32 @@ const style = {
 } as const
 
 /** Mock svg Size */
-const DEFAULT_SIZE = { width: 500, height: 500 }
+const DEFAULT_SIZE = { width: 700, height: 700 }
 
-function SvgContourContainer(): JSX.Element {
+function PolygonDrawerContainer(): JSX.Element {
   const { loadingState, image } = useImage({
     wadouri: IMAGES[12],
   })
   const { viewport, setViewport } = useViewport()
-  const { contours, focusedContour } = useContour<Contour>({
-    mode: 'contour',
-    initalContours: CONTOURS,
-  })
+  const {
+    contours,
+    focusedContour,
+    addContour,
+    removeContour,
+    focusContour,
+    removeAllContours,
+  } = useContour<Contour>({ mode: 'polygon' })
 
   return (
     <>
+      <Button
+        data-cy-name="remove-button"
+        marginBottom="10px"
+        colorScheme="blue"
+        onClick={removeAllContours}
+      >
+        remove all
+      </Button>
       <Box data-cy-loaded={loadingState}>
         <Resizable style={style} defaultSize={DEFAULT_SIZE}>
           <InsightViewer
@@ -44,23 +53,25 @@ function SvgContourContainer(): JSX.Element {
             onViewportChange={setViewport}
           >
             {loadingState === 'success' && (
-              <SvgContourViewer
+              <Annotation
                 focusedContour={focusedContour}
                 width={DEFAULT_SIZE.width}
                 height={DEFAULT_SIZE.height}
                 contours={contours}
                 polygonAttrs={getPolygonStyles}
                 showPolygonLabel
+                onAdd={addContour}
+                onFocus={focusContour}
+                onRemove={removeContour}
+                isDrawing
+                mode="polygon" // If no mode is defined, the default value is polygon.
               />
             )}
           </InsightViewer>
         </Resizable>
       </Box>
-      <Box>
-        <CodeBlock code={CODE} />
-      </Box>
     </>
   )
 }
 
-export default SvgContourContainer
+export default PolygonDrawerContainer
